requirements.txt
Flask
pdf2image
pillow
pytesseract
